elmo-requirements.txt
absl-py==0.9.0
aio-pika==6.4.1
aiormq==3.2.2
astor==0.8.1
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Cython==0.29.14
DAWG-Python==0.7.2
decorator==4.4.2
deeppavlov==0.11.0
docopt==0.6.2
fastapi==0.47.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.30.0
h11==0.9.0
h5py==2.10.0
httptools==0.0.13
idna==2.8
importlib-metadata==1.7.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
joblib==0.16.0
jsonpickle==1.4.1
jupyter-client==6.1.5
jupyter-core==4.6.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
llvmlite==0.33.0
Markdown==3.2.2
multidict==4.7.6
nltk==3.4.5
numba==0.50.1
numpy==1.18.0
opt-einsum==3.2.1
overrides==2.7.0
pamqp==2.3.0
pandas==0.25.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pycparser==2.20
pydantic==1.3
Pygments==2.6.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942
pymystem3==0.2.0
pyOpenSSL==19.1.0
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.1
pytz==2019.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.22.0
ruamel.yaml==0.15.100
rusenttokenize==0.0.5
russian-tagsets==0.6
sacremoses==0.0.35
scikit-learn==0.21.2
scipy==1.4.1
six==1.15.0
starlette==0.12.9
tensorboard==1.15.0
tensorflow==1.15.3
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.3
tensorflow-hub==0.8.0
termcolor==1.1.0
tornado==6.0.4
tqdm==4.41.1
traitlets==4.3.3
umap-learn==0.4.6
urllib3==1.25.9
uvicorn==0.11.1
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
